added "MapCoverTotal()" function to spatial.R
It is similar to "MapCoverChange()" but displays total understory cover as opposed to understory change.

Also added a "cycle_filter" argument to UnderNativityCover.plot.nat_v_non to remove filter toggle checkbox when crosstalking with maps (which currently don't allow more than one cycle to be displayed at a time).